requirements.txt
# Core web framework
fastapi==0.115.14
uvicorn[standard]==0.35.0

# OpenAI API & Tokenizer
openai==1.93.0
tiktoken==0.9.0

# HTTP Client
httpx==0.28.1

# Qdrant vector DB client
qdrant-client==1.14.3

# LangChain & modules
langchain==0.3.26
langchain-core==0.3.67
langchain-community==0.3.26
langchain-text-splitters==0.3.8

# Token counting
tiktoken==0.9.0

# PDF parsing
PyMuPDF==1.26.1

# Utils
python-dotenv==1.1.1
tqdm==4.67.1
